move out Embedding to separate file
