+++ metaseq/models/transformer_lm.py
@@ -22,9 +22,9 @@ from metaseq.models import (
 )
 from metaseq.models.transformer import (
     DEFAULT_MIN_PARAMS_TO_WRAP,
-    Embedding,
     TransformerDecoder,
 )
+from metaseq.models.embedding import Embedding
 
 DEFAULT_MAX_TARGET_POSITIONS = 1024
